moved Settings.h to the tvheadend/ subdirectory to reflect its actual namespace
